multi-day event conversion missing all day events
The new GoogleCalendar::EventConversion service will convert a
multi-day event to an array of events. The implementation to this point
is missing the middle days of the date range. As of this commit it
creates only the first and last events in the range with the proper
start and end times.