Handle all day events from google calendar (#36)
* add service response

* multi-day event conversion missing all day events

The new GoogleCalendar::EventConversion service will convert a
multi-day event to an array of events. The implementation to this point
is missing the middle days of the date range. As of this commit it
creates only the first and last events in the range with the proper
start and end times.

* tweak layout spacing and colors

* add event conversion to handle multi-day google calendar events

* add all day events

* tell calendar what range to build instead of assuming

* make sure all day events show up on the right date